Pass around typmod as int16.

+++ src/backend/access/common/printtup.c
@@ -124,7 +124,7 @@ printtup(HeapTuple tuple, TupleDesc typeinfo)
 		{
 			outputstr = fmgr(typoutput, attr,
 							 gettypelem(typeinfo->attrs[i]->atttypid),
-										(int)typeinfo->attrs[i]->atttypmod);
+										typeinfo->attrs[i]->atttypmod);
 			pq_putint(strlen(outputstr) + VARHDRSZ, VARHDRSZ);
 			pq_putnchar(outputstr, strlen(outputstr));
 			pfree(outputstr);
@@ -191,7 +191,7 @@ debugtup(HeapTuple tuple, TupleDesc typeinfo)
 		{
 			value = fmgr(typoutput, attr,
 						 gettypelem(typeinfo->attrs[i]->atttypid),
-							   (int)typeinfo->attrs[i]->atttypmod);
+									typeinfo->attrs[i]->atttypmod);
 			printatt((unsigned) i + 1, typeinfo->attrs[i], value);
 			pfree(value);
 		}
